// graphql-types.ts
export type Maybe<T> = T | null;
// ...
export type MarkdownRemark = Node & {
  id: Scalars['ID'];
  frontmatter?: Maybe<MarkdownRemarkFrontmatter>;
  excerpt?: Maybe<Scalars['String']>;
  rawMarkdownBody?: Maybe<Scalars['String']>;
  fileAbsolutePath?: Maybe<Scalars['String']>;
  fields?: Maybe<MarkdownRemarkFields>;
  html?: Maybe<Scalars['String']>;
  htmlAst?: Maybe<Scalars['JSON']>;
  excerptAst?: Maybe<Scalars['JSON']>;
  headings?: Maybe<Array<Maybe<MarkdownHeading>>>;
  timeToRead?: Maybe<Scalars['Int']>;
  tableOfContents?: Maybe<Scalars['String']>;
  wordCount?: Maybe<MarkdownWordCount>;
  parent?: Maybe<Node>;
  children: Array<Node>;
  internal: Internal;
};
// ...
export type MarkdownRemarkFields = {
  slug?: Maybe<Scalars['String']>;
};
// ...
export type MarkdownRemarkFrontmatter = {
  title?: Maybe<Scalars['String']>;
  date?: Maybe<Scalars['Date']>;
  previewImage?: Maybe<File>;
  image?: Maybe<File>;
  preview?: Maybe<Scalars['String']>;
};
// ...
export type PostPreviewImageFragment = { fluid?: Maybe<GatsbyImageSharpFluid_NoBase64Fragment> };

export type PostImageFragment = { fluid?: Maybe<GatsbyImageSharpFluid_NoBase64Fragment> };

export type PostFragment = (
  Pick<MarkdownRemark, 'id'>
  & { content: MarkdownRemark['html'], preview: MarkdownRemark['excerpt'] }
  & { frontmatter?: Maybe<(
    Pick<MarkdownRemarkFrontmatter, 'title' | 'date'>
    & { previewImage?: Maybe<{ childImageSharp?: Maybe<PostPreviewImageFragment> }>, image?: Maybe<{ childImageSharp?: Maybe<PostImageFragment> }> }
  )>, fields?: Maybe<Pick<MarkdownRemarkFields, 'slug'>> }
);
// ...
export type PostPageQuery = { markdownRemark?: Maybe<PostFragment> };